Support dependency injection for SAML strategies
